backend/requirements.txt
flask
flask-cors
tensorflow
numpy
pillow
pymongo
scipy